New testing design (#590)
* VIEWS FINALLY WORK

In this commit which had hours leading up to it I have finally figured out a consistent way to specify how to respond to a view in tests. Awesome. However one Card test failed in one of many trials, so not perfect

* Much improve testing ecosystem, add explanation

* Test design overhaul complete

All tests now function in the new testing system

* Improve testing info file

* Make formatting better